+++ packages/repack/src/utils/index.ts
@@ -4,5 +4,5 @@ export * from './getDirname.js';
 export * from './getPublicPath.js';
 export * from './getResolveOptions.js';
 export * from './getModulePaths.js';
-export * from './getJavaScriptTransformRules.js';
+export * from './getJsTransformRules.js';
 export * from './getSwcLoaderOptions.js';
